add name to success alert
